added swagger ui for api documentation

+++ src/main/java/io/elearning/data/repository/CourseRepository.java
@@ -4,7 +4,6 @@ import io.elearning.data.models.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,5 +11,5 @@ import java.util.Optional;
 public interface CourseRepository extends JpaRepository<Course, Long> {
     Optional<Course> findByCourseName(String courseName);
     Optional<Course> findAllByCourseName(String courseName);
-    List<Course> findAllByCoursePrice(BigDecimal coursePrice);
+    List<Course> findAllByCoursePrice(double coursePrice);
 }
