+++ src/main/java/io/elearning/web/service/CourseService.java
@@ -5,7 +5,6 @@ import io.elearning.data.models.Course;
 import io.elearning.exceptions.CourseException;
 import io.elearning.exceptions.UserException;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface CourseService {
@@ -14,7 +13,7 @@ public interface CourseService {
     Course updateCourse(CourseDto courseDto, Long courseId) throws CourseException;
     List<Course> getAllCourses();
     Course getACourseByName(String courseName) throws CourseException;
-    List<Course> getCoursesByPrice(BigDecimal price);
+    List<Course> getCoursesByPrice(double price);
     void deleteACourse(Long courseId) throws CourseException;
     Course getACourseById(Long courseId) throws CourseException;
 }
